Make lattice_size a float instead of uint64, which saves a couple extra registers on Fermi.

+++ branches/specialq-nobatch/gnfs/poly/stage1/stage1_core_gpu/stage1_core_nosq.cu
@@ -56,7 +56,7 @@ sieve_kernel_48(q_soa_t *qbatch,
 			uint64 p2 = wide_sqr32(p);
 			uint32 pinvmodq = modinv32(p, q);
 
-			uint64 lattice_size = curr_p->lattice_size;
+			float lattice_size = curr_p->lattice_size;
 			uint32 num_proots = curr_p->num_roots;
 			uint64 pinv, tmp;
 
@@ -123,7 +123,7 @@ sieve_kernel_64(q_soa_t *qbatch,
 			uint64 p2 = wide_sqr32(p);
 			uint32 pinvmodq = modinv32(p, q);
 
-			uint64 lattice_size = curr_p->lattice_size;
+			float lattice_size = curr_p->lattice_size;
 			uint32 num_proots = curr_p->num_roots;
 			uint64 pinv, tmp;
